Moving ListingTitleLink component -> TS
