+++ app/assets/img/nav/ic-search.png
