--- app/components/layout/index.js
@@ -1,2 +1,0 @@
-export { default as ServiceCard } from './ServiceCard';
-export { default as ListingTitleLink } from './ListingTitleLink';
