+++ app/components/listing/ServiceCard.tsx
@@ -1,0 +1,16 @@
+import React from 'react';
+import { Link } from 'react-router-dom';
+import { Service } from '../../models';
+// import StreetViewImage from 'components/maps/StreetViewImage';
+
+export const ServiceCard = ({ service }: { service: Service }) => {
+  const { id, name, long_description } = service;
+  return (
+    <Link to={{ pathname: `/services/${id}` }} className="card">
+      <header className="content">
+        <h3>{ name }</h3>
+        <p>{ long_description }</p>
+      </header>
+    </Link>
+  );
+};
