+++ package.json
@@ -11,7 +11,6 @@
     "@babel/polyfill": "^7.4.4",
     "@sentry/browser": "^4.0.6",
     "@typeform/embed": "^0.13.1",
-    "@types/react-modal": "^3.12.0",
     "classnames": "^2.2.6",
     "connected-react-router": "^6.5.2",
     "copy-webpack-plugin": "^5.1.1",
@@ -22,6 +21,7 @@
     "moment": "^2.24.0",
     "prop-types": "^15.5.10",
     "qs": "^6.5.1",
+    "react": "^16.8.6",
     "react-burger-menu": "^2.6.0",
     "react-dom": "^16.8.6",
     "react-ga": "^2.4.1",
@@ -33,15 +33,14 @@
     "react-intercom": "^1.0.15",
     "react-markdown": "^3.3.2",
     "react-modal": "^3.12.1",
-    "react-redux-form": "^1.12.1",
     "react-redux": "^6.0.1",
+    "react-redux-form": "^1.12.1",
     "react-router-dom": "^5.1.2",
     "react-select": "^1.0.0-rc.5",
     "react-tippy": "^1.2.2",
-    "react": "^16.8.6",
+    "redux": "^3.6.0",
     "redux-immutable-state-invariant": "^1.2.4",
     "redux-thunk": "^2.1.0",
-    "redux": "^3.6.0",
     "url-search-params-polyfill": "^7.0.0",
     "whatwg-fetch": "^2.0.1"
   },
@@ -54,11 +53,13 @@
     "@babel/preset-react": "^7.0.0",
     "@babel/preset-typescript": "^7.13.0",
     "@babel/register": "^7.4.4",
+    "@types/enzyme": "^3.10.9",
     "@types/qs": "^6.9.6",
-    "@types/react": "^17.0.3",
     "@types/react-dom": "^17.0.3",
+    "@types/react-modal": "^3.12.0",
     "@types/react-redux": "^7.1.16",
     "@types/react-router-dom": "^5.1.7",
+    "@types/react": "^17.0.3",
     "@types/redux": "^3.6.0",
     "@typescript-eslint/eslint-plugin": "^4.14.1",
     "@typescript-eslint/parser": "^4.14.1",
